[GTK] Move GtkUtilities and GtkVersioning from WebCore to WebKit https://bugs.webkit.org/show_bug.cgi?id=300781
Reviewed by Patrick Griffis.

They are only used in WebKit.

Canonical link: https://commits.webkit.org/301544@main
