+++ Source/WebCore/SourcesGTK.txt
@@ -84,7 +84,6 @@ platform/graphics/glib/ImageAdapterGLib.cpp
 
 platform/graphics/gtk/ColorGtk.cpp
 platform/graphics/gtk/GdkCairoUtilities.cpp
-platform/graphics/gtk/GdkSkiaUtilities.cpp
 platform/graphics/gtk/IconGtk.cpp
 platform/graphics/gtk/SystemFontDatabaseGTK.cpp
 
@@ -93,7 +92,6 @@ platform/graphics/opentype/OpenTypeVerticalData.cpp
 platform/graphics/x11/XErrorTrapper.cpp @no-unify
 
 platform/gtk/GRefPtrGtk.cpp
-platform/gtk/GtkUtilities.cpp @no-unify
 platform/gtk/LocalizedStringsGtk.cpp
 platform/gtk/PasteboardGtk.cpp
 platform/gtk/PlatformKeyboardEventGtk.cpp
