+++ Source/WebCore/platform/graphics/gtk/GdkCairoUtilities.cpp
@@ -32,9 +32,6 @@
 #include "IntSize.h"
 #include <cairo.h>
 #include <gtk/gtk.h>
-#include <mutex>
-#include <wtf/NeverDestroyed.h>
-#include <wtf/glib/GUniquePtr.h>
 
 namespace WebCore {
 
@@ -43,23 +40,6 @@ GRefPtr<GdkPixbuf> cairoSurfaceToGdkPixbuf(cairo_surface_t* surface)
     IntSize size = cairoSurfaceSize(surface);
     return adoptGRef(gdk_pixbuf_get_from_surface(surface, 0, 0, size.width(), size.height()));
 }
-
-#if USE(GTK4)
-GRefPtr<GdkTexture> cairoSurfaceToGdkTexture(cairo_surface_t* surface)
-{
-    ASSERT(cairo_image_surface_get_format(surface) == CAIRO_FORMAT_ARGB32);
-    auto width = cairo_image_surface_get_width(surface);
-    auto height = cairo_image_surface_get_height(surface);
-    if (width <= 0 || height <= 0)
-        return nullptr;
-    auto stride = cairo_image_surface_get_stride(surface);
-    auto* data = cairo_image_surface_get_data(surface);
-    GRefPtr<GBytes> bytes = adoptGRef(g_bytes_new_with_free_func(data, height * stride, [](gpointer data) {
-        cairo_surface_destroy(static_cast<cairo_surface_t*>(data));
-    }, cairo_surface_reference(surface)));
-    return adoptGRef(gdk_memory_texture_new(width, height, GDK_MEMORY_DEFAULT, bytes.get(), stride));
-}
-#endif
 
 } // namespace WebCore
