+++ Source/WebCore/platform/graphics/gtk/GdkCairoUtilities.h
@@ -29,16 +29,9 @@
 
 #include <wtf/glib/GRefPtr.h>
 
-#if USE(GTK4)
-typedef struct _GdkTexture GdkTexture;
-#endif
-
 namespace WebCore {
 
 GRefPtr<GdkPixbuf> cairoSurfaceToGdkPixbuf(cairo_surface_t*);
-#if USE(GTK4)
-GRefPtr<GdkTexture> cairoSurfaceToGdkTexture(cairo_surface_t*);
-#endif
 }
 
 #endif // USE(CAIRO)
