--- Source/WebCore/platform/graphics/gtk/GdkSkiaUtilities.h
@@ -1,52 +1,0 @@
-/*
- * Copyright (C) 2024 Igalia S.L.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- * 1. Redistributions of source code must retain the above copyright
- *    notice, this list of conditions and the following disclaimer.
- * 2. Redistributions in binary form must reproduce the above copyright
- *    notice, this list of conditions and the following disclaimer in the
- *    documentation and/or other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY APPLE INC. ``AS IS'' AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
- * IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
- * PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL APPLE INC. OR
- * CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
- * EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
- * PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
- * PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
- * OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
- * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-#pragma once
-
-#if USE(SKIA)
-
-#include <gdk/gdk.h>
-WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_BEGIN
-#include <skia/core/SkImage.h>
-WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_END
-#include <wtf/glib/GRefPtr.h>
-
-#if !USE(GTK4)
-#include <cairo.h>
-#endif
-
-namespace WebCore {
-
-#if USE(GTK4)
-GRefPtr<GdkTexture> skiaImageToGdkTexture(SkImage&);
-#else
-RefPtr<cairo_surface_t> skiaImageToCairoSurface(SkImage&);
-#endif
-
-GRefPtr<GdkPixbuf> skiaImageToGdkPixbuf(SkImage&);
-
-}
-
-#endif // USE(SKIA)
