+++ Source/WebKit/Shared/gtk/NativeWebTouchEventGtk.cpp
@@ -28,9 +28,8 @@
 
 #if ENABLE(TOUCH_EVENTS)
 
-#include <WebCore/GtkVersioning.h>
+#include "GtkVersioning.h"
 #include "WebEventFactory.h"
-#include <gdk/gdk.h>
 
 namespace WebKit {
