+++ Source/WebKit/UIProcess/Notifications/glib/NotificationService.cpp
@@ -60,7 +60,7 @@ WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_BEGIN
 WTF_IGNORE_WARNINGS_IN_THIRD_PARTY_CODE_END
 
 #if PLATFORM(GTK)
-#include <WebCore/GtkVersioning.h>
+#include "GtkVersioning.h"
 #endif
 
 #if HAVE(GDESKTOPAPPINFO)
