+++ Source/WebKit/UIProcess/gtk/WebDataListSuggestionsDropdownGtk.cpp
@@ -26,11 +26,11 @@
 #include "config.h"
 #include "WebDataListSuggestionsDropdownGtk.h"
 
+#include "GtkUtilities.h"
+#include "GtkVersioning.h"
 #include "WebPageProxy.h"
 #include <WebCore/DataListSuggestionInformation.h>
 #include <WebCore/GUniquePtrGtk.h>
-#include <WebCore/GtkUtilities.h>
-#include <WebCore/GtkVersioning.h>
 #include <WebCore/IntPoint.h>
 #include <wtf/glib/GRefPtr.h>
 #include <wtf/glib/GUniquePtr.h>
@@ -164,7 +164,7 @@ void WebDataListSuggestionsDropdownGtk::show(WebCore::DataListSuggestionInformat
     auto* monitor = gdk_display_get_monitor_at_window(display, gtk_widget_get_window(m_webView));
 #endif
     GdkRectangle area;
-    WebCore::monitorWorkArea(monitor, &area);
+    monitorWorkArea(monitor, &area);
     int width = std::min(information.elementRect.width(), area.width);
     size_t itemCount = std::min<size_t>(information.suggestions.size(), (area.height / 3) / itemHeight);
