+++ Tools/Scripts/webkitpy/style/checker.py
@@ -548,11 +548,11 @@ _SKIPPED_FILES_WITH_WARNING = [
     os.path.join('Source', 'WebKit', 'WebProcess', 'InjectedBundle', 'API', 'gtk', 'DOM'),
 
     os.path.join('Source', 'JavaScriptCore', 'API', 'glib', 'jsc.h'),
-    os.path.join('Source', 'WebCore', 'platform', 'gtk', 'GtkVersioning.h'),
     os.path.join('Source', 'WebCore', 'platform', 'graphics', 'gbm', 'GBMVersioning.h'),
     os.path.join('Source', 'WebKit', 'UIProcess', 'API', 'gtk', 'webkit2.h'),
     os.path.join('Source', 'WebKit', 'UIProcess', 'API', 'gtk', 'webkit.h'),
     os.path.join('Source', 'WebKit', 'UIProcess', 'API', 'wpe', 'webkit.h'),
+    os.path.join('Source', 'WebKit', 'UIProcess', 'gtk', 'GtkVersioning.h'),
     os.path.join('Source', 'WebKit', 'WPEPlatform', 'wpe', 'wpe-platform.h'),
     os.path.join('Source', 'WebKit', 'WPEPlatform', 'wpe', 'drm', 'wpe-drm.h'),
     os.path.join('Source', 'WebKit', 'WPEPlatform', 'wpe', 'headless', 'wpe-headless.h'),
