+++ Tools/TestWebKitAPI/Tests/WebKitGtk/TestPrinting.cpp
@@ -19,7 +19,7 @@
 
 #include "config.h"
 #include "WebViewTest.h"
-#include <WebCore/GtkVersioning.h>
+#include <WebKit/GtkVersioning.h>
 #include <glib/gstdio.h>
 #include <wtf/glib/GRefPtr.h>
