+++ Tools/TestWebKitAPI/glib/WebKitGLib/gtk/WebViewTestGtk.cpp
@@ -22,7 +22,7 @@
 #include "WebViewTest.h"
 
 #include <WebCore/GUniquePtrGtk.h>
-#include <WebCore/GtkVersioning.h>
+#include <WebKit/GtkVersioning.h>
 #include <gtk/gtk.h>
 #include <webkit/WebKitWebViewBaseInternal.h>
