+++ Tools/TestWebKitAPI/gtk/PlatformWebViewGtk.cpp
@@ -27,7 +27,7 @@
 #include "PlatformWebView.h"
 
 #include <WebCore/GUniquePtrGtk.h>
-#include <WebCore/GtkVersioning.h>
+#include <WebKit/GtkVersioning.h>
 #include <WebKit/WKRetainPtr.h>
 #include <WebKit/WKView.h>
 #include <gtk/gtk.h>
